backend/requirements.txt
groq
langgraph
python-dotenv
tavily-python
pytrends
requests
newsapi-python